Terminado la logica de los grupos, faltan los permisos

+++ app/controllers/groups_controller.rb
@@ -1,3 +1,4 @@
+# -*- coding: utf-8 -*-
 class GroupsController < ApplicationController
 
   # GET /groups/1
@@ -38,18 +39,49 @@ class GroupsController < ApplicationController
   # PUT /groups/1
   # PUT /groups/1.json
   def update
-    MembersInGroup.create(:user_id => params[:user_id],:group_id => params[:group_id])
-    redirect_to :new_group
+    user = User.find_by_personal_url(params[:personal_url])
+    if current_user.id != user.id then
+      redirect_to show_user(user.personal_url)
+      return
+    end
+    group = Group.find_by_id(params[:id])
+    group.update_attributes(:name => params[:name], :description => params[:description])
+    redirect_to :show_groups
   end
 
   # DELETE /groups/1
   # DELETE /groups/1.json
   def destroy
+    user = User.find_by_personal_url(params[:personal_url])
+    if current_user.id != user.id
+      redirect_to show_user(user)
+      return
+    end
     group = Group.find_by_id(params[:group_id])
     group.members_in_group.each do |member|
       member.destroy
     end
     group.delete
-    redirect_to :new_group
+    redirect_to :show_groups
+  end
+
+  # GET /groups/edit
+  # GET /gropus/1.json
+  def edit
+    @user = User.find_by_personal_url(params[:personal_url])
+    if @user.id != current_user.id then
+      redirect_to :root_path
+      return
+    else
+      @group = Group.find_by_id(params[:id])
+      if @group.nil? then
+        # Si el grupo no existe, entonces no se puede editar
+        redirect_to :show_user
+        return
+      end
+      # Si no eres el dueño del grupo no lo puedes editar
+      redirect_to :show_user if @group.user_id != @user.id
+      return
+    end
   end
 end
